requirements.txt
accelerate==0.34.2
aiohttp==3.10.8
attrs==24.2.0
backoff==2.2.1
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
langchain==0.3.1
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-google-genai==2.0.0
numpy==1.26.4
pydantic==2.9.2
torch==2.4.1
requests
flask
Flask-HTTPAuth
supabase
APScheduler